Sanitize audio params, fix types & tests
Add numeric sanitizers and helpers, tighten types and casts, and update tests.

- Introduce sanitize_delay_samples and reverb_ir_size_samples in resonance_math.h to handle NaN/Inf and compute IR sizes safely; include <cstdint>.
- Use the new sanitizers for reflection params and when fetching reverb params to avoid Steam Audio NaN/Inf/delay issues.
- Replace several raw int/float operations with explicit casts and safer math (std::sqrt, lroundf) to prevent narrowing and platform issues.
- Harden file/buffer and index handling (file length, triangle index counts, multimesh instance count, curve sample sizes) using int64/int32 casts.
- Small logic fixes: initialize mix channel variables, consolidate attenuation boundary checks, and avoid accidental integer/float truncation.
- Rename and expose helper/callback identifiers (find_resonance_config_in_tree, distance_attenuation_callback) and rename the singleton variable to g_resonance_server_singleton for clarity.
- Update unit tests: add BakeRunner pathing-hash test ensuring parity with C++ dict format and adjust Android realtime rays test to reflect passthrough behavior.

These changes improve numeric robustness, cross-language consistency, and test coverage to prevent runtime warnings and platform-dependent bugs.

+++ src/resonance_processor_direct.cpp
@@ -2,6 +2,7 @@
 #include "resonance_log.h"
 #include "resonance_server.h"
 #include "resonance_utils.h"
+#include <cmath>
 #include <cstring>
 
 namespace godot {
@@ -257,7 +258,7 @@ void ResonanceDirectProcessor::process(
     IPLVector3 local_dir = {x, y, z};
     float len_sq = x * x + y * y + z * z;
     if (len_sq > resonance::kDegenerateVectorEpsilonSq) {
-        float len = sqrt(len_sq);
+        float len = static_cast<float>(std::sqrt(static_cast<double>(len_sq)));
         local_dir.x /= len;
         local_dir.y /= len;
         local_dir.z /= len;
